Binary searched length cache and no superfluous cache clearing.
LengthTo showed up as the main hot-spot on profiles of large method emissions.  The replaced implementation would always start at the beginning in the event of a miss (even though length determination might have been done through a higher offset, and the cache was being cleared even in the most common event of a simply appended instruction.

Closes PR #5

Co-authored-by: n8allan <nate@digithought.com>